[FIX] remove hack to avoid IntegrityError
* fix cause of IntegrityError and remove hack
* modify the test to cover this

this IntegrityError was due to the fact that the same dict was re-used to
create multiple timesheets. during the creation, the provided dict is
modified, adding a default follower. this resulted in the same follower
being added multiple times (one for each loop iteration), and thus in
trying to create multiple follower records with the same values.